requirements.txt
web.py==0.62
psycopg2-binary==2.9.3
